[NUI][Samples] Added RiveTest Sample.
This sample was created for test multiple rive files loaded from the
device storage. It searches for files in the application resources
directory. Then creates a list of the available .riv files. To select
a file, point to it in the scrollable list. The selected file is
highlighted in yellow. By default, no file is selected. In case of any
issue error message is displayed.